Relocate to table submodule; code and comment improvements
